+++ toolchain/internal/asm/flags_test.go
@@ -11,15 +11,15 @@ func TestFlags(t *testing.T) {
 
 	enc, err := fs.Encode("ab")
 	assert.NoError(t, err)
-	assert.Equal(t, int64(0b110000), enc)
+	assert.Equal(t, uint32(0b110000), enc)
 
 	enc, err = fs.Encode("a")
 	assert.NoError(t, err)
-	assert.Equal(t, int64(0b100000), enc)
+	assert.Equal(t, uint32(0b100000), enc)
 
 	enc, err = fs.Encode("c")
 	assert.NoError(t, err)
-	assert.Equal(t, int64(0b001000), enc)
+	assert.Equal(t, uint32(0b001000), enc)
 
 	enc, err = fs.Encode("d")
 	assert.Error(t, err)
